WIP Address a part of review comments from @tdas
* TODO list
  * replace all the usages for direct call of store.xxx whenever state manager is available
  * add iterator / remove in StreamingAggregationStateManager to remove restoreOriginRow
  * add docs
